fix sonour cloud issue

+++ source/core/src/main/com/csse3200/game/areas/GameArea.java
@@ -1183,6 +1183,6 @@ public abstract class GameArea implements Disposable {
                             Vector2 camPos) {
     }
     public float roomNumber() {
-        return (float)getRoomNumber();
+        return getRoomNumber();
     }
 }
